+++ instrumentation-api/src/test/groovy/io/opentelemetry/instrumentation/api/tracer/NetPeerUtilsTest.groovy
@@ -17,6 +17,7 @@
 package io.opentelemetry.instrumentation.api.tracer
 
 import io.opentelemetry.auto.test.utils.ConfigUtils
+import io.opentelemetry.instrumentation.api.tracer.utils.NetPeerUtils
 import io.opentelemetry.trace.attributes.SemanticAttributes
 
 class NetPeerUtilsTest extends BaseTracerTest {
