+++ instrumentation/elasticsearch/elasticsearch-transport-5.0/src/main/java/io/opentelemetry/instrumentation/auto/elasticsearch/transport/v5_0/TransportActionListener.java
@@ -19,7 +19,7 @@ package io.opentelemetry.instrumentation.auto.elasticsearch.transport.v5_0;
 import static io.opentelemetry.instrumentation.auto.elasticsearch.transport.ElasticsearchTransportClientDecorator.DECORATE;
 
 import com.google.common.base.Joiner;
-import io.opentelemetry.instrumentation.api.tracer.NetPeerUtils;
+import io.opentelemetry.instrumentation.api.tracer.utils.NetPeerUtils;
 import io.opentelemetry.trace.Span;
 import io.opentelemetry.trace.attributes.SemanticAttributes;
 import org.elasticsearch.action.ActionListener;
